Mock objects of project
Mock objects of project that show rough API of components of whole application.
Currently, real-time processing components for backend, frontend or AI interviewer are missing, because it is a complex task to construct such system.
Probably, after small research on this topic, I will make prototype for real-time components as well.