Evaluation page
Evaluation page is now almost done - there is a matter of optimization to consider, since currently we try to gather too much of job requirements so that it takes a long time, and, more importantly, when combined with scorings on those requirements it exceeds allowed cookie size! Need to optimize that and to use `clear_evaluation` function somewhere.